+++ tests/test_swagger.py
@@ -185,6 +185,49 @@ class ExtractContentTypesTest(unittest.TestCase):
         self.assertEquals(spec['paths']['/icecream/{flavour}']['put']['consumes'],
                           ['application/json', 'text/xml'])
 
+    def test_single_ctype_callable(self):
+
+        service = Service("IceCream", "/icecream/{flavour}")
+
+        def my_ctype_callable(request):
+            return 'application/octet-stream'
+
+        class IceCream(object):
+            @service.put(content_type=my_ctype_callable)
+            def view_put(self, request):
+                return self.request.validated
+
+        swagger = CorniceSwagger([service])
+        spec = swagger.generate()
+        self.assertEquals(spec['paths']['/icecream/{flavour}']['put']['consumes'],
+                          [])
+
+    def test_mixed_ctype_string_and_callable(self):
+
+        service = Service("IceCream", "/icecream/{flavour}")
+
+        def my_ctype_callable(request):
+            return 'application/octet-stream'
+
+        def my_ctype_callable_2(request):
+            return 'image/png'
+
+        class IceCream(object):
+            @service.put(content_type=(
+                    my_ctype_callable,
+                    'application/json',
+                    my_ctype_callable_2,
+                    'image/jpeg'
+                    )
+            )
+            def view_put(self, request):
+                return self.request.validated
+
+        swagger = CorniceSwagger([service])
+        spec = swagger.generate()
+        self.assertEquals(spec['paths']['/icecream/{flavour}']['put']['consumes'],
+                          ['application/json', 'image/jpeg'])
+
     def test_ignore_multiple_views_by_ctype(self):
 
         service = Service("IceCream", "/icecream/{flavour}")